setup_DTO : DTO, Controller, Service 생성 및 기본 세팅

+++ src/main/java/com/example/crud_backend/domain/Users.java
@@ -16,12 +16,12 @@ public class Users {
     @Column(length=1, nullable=false)
     private String sex;
     @Column(nullable=false)
-    private Long age;
+    private int age;
     private String phone;
     private String email;
 
     @Builder
-    public Users (String id, String name, String pw, String sex, Long age, String phone, String email) {
+    public Users (String id, String name, String pw, String sex, int age, String phone, String email) {
         this.id=id;
         this.name=name;
         this.pw=pw;
